README.md: Update build instructions
The 'make binaries' command requires the existence of
libc.a.  On Fedora this file is included in the glibc-static
package.  If this is not installed the user will get the error
'/usr/bin/ld: cannot find -lc'.  On Ubuntu this is not an issue.
The README was updated to reflect this dependency.

Signed-off-by: Ryan Cole <rcyoalne@gmail.com>