Makefile: fix remove docs in make clean
On Ubuntu, `/bin/sh` is used in Makefile by default, so
a.{1,2,3} isn't evaluated to a.1 a.2 a.3, another solution
is to use `SHELL=/bin/bash` in Makefile, but I don't think
we should bind to bash this early, at least not for this
case.

This also remove the removal of docs/*.1 which doesn't
exist for now.

Signed-off-by: Qiang Huang <h.huangqiang@huawei.com>