server: container_status: return image name if available
If we create a container using the image ID like
771cd5947d5ea4bf8e8f4900dd357dbb67e7b16486c270f8274087d182d457c6, then
a call to container_status will return that same ID for the "Image"
field in ContainerStatusResponse.

This patch matches dockershim behavior and return the first tagged name
if available from the image store.

This is also needed to fix a failure in k8s e2d tests.

Reference:
https://github.com/kubernetes/kubernetes/pull/39298/files#diff-c7dd39479fd733354254e70845075db5R369
Reference:
https://github.com/kubernetes/kubernetes/blob/67a5bf8454bb839ee80a4245e9e09fc8f700fa84/test/e2e/framework/util.go#L1941

Signed-off-by: Antonio Murdaca <runcom@redhat.com>