server: Don't stop a sandbox or a container in "created" state
There is no point in trying to stop a container that is not currently
running. More than that, the OCI specification is very clear that
trying to issue a "kill" command on a container which is not running
must be considered as an error.

Refer to OCI specification here:
https://github.com/opencontainers/runtime-spec/blob/master/runtime.md#kill

In case of a pod sandbox, OCI is not forcing anything, but I think it
does make sense not to try to stop a pod which is not running. It could
lead to unexpected behaviors, and thinking about semantics, we should
not be able to stop something which has not been started first.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>