test: Make the storage driver configurable for bats
It is currently impossible to start bats with a different storage
driver than the default one. This is because helpers.bash does not
rely on crio.conf file that we can find at /etc/crio/crio.conf.

This patch introduces a new environment variable called
"CRIO_STORAGE_DRIVER" to let the caller of the bats decide if he
wants a different storagedriver when running its tests.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>