Return image references in ImageStatus()
The image's canonical reference is its repository name paired with a
digest of the image's manifest instead of a tag, so compute and return
that value as the image's reference in ImageStatus() and in
ContainerStatus().

We don't auto-store a name based on the image digest when we pull one by
tag, but then CRI doesn't need us to do that.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>