I would like to add @nalind to the OWNERS of cri-o
He is a key contributor to the CRI-O project, he has contributed a lot to cri-o
and the projects/libraries

He is an expert in containers/storage and really solid on containers/image
key components of the cri-o project.  We need to get his advise and appoval
on lots of pull requests interacting with the storage and image layers.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>