tests: Support IRC notifications
Add an 'irc_channel_notice' role to help post messages in a somewhat
generic way.  It verifies it's own required variables and is able
to translate between commit-author names and irc nicks.  The message
itself is generated via a jinja2 template, making it easy to swap in
alternate messages for different situations.

The intent is this role is applied during different stages of the
crio-integration-playbook.  As long as at least one variable differs
between references (e.g. verb), it will execute multiple times.

Test with:

- hosts: localhost
  gather_subset: network
  roles:
    - role: irc_channel_notice
      verb: testing
      commit_author: willgates
      pullrequest: 889

Signed-off-by: Chris Evich <cevich@redhat.com>