server: Do not suppress image repull in cri-o
The PullImage function is trying to be too smart and as a result
suppressing pulls where the imagePullPolicy is Always. This policy
is not plumbed down and so we must trust the upper layers to do
this check. This matches the behavior of docker, which does repull
if the tag has changed. We must rely on the lower layers to not
redownload an image already in the cache.

Signed-off-by: Andrew Pilloud <andrewpilloud@igneoussystems.com>